requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0

# Data science
numpy<2.0
pandas>=1.3.0

# Progress bars and utilities
tqdm>=4.65.0

# Computer vision models
timm>=0.9.2

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# System monitoring
psutil>=5.8.0

# Model complexity analysis
thop>=0.1.1.post2209072238

# Plotting
seaborn>=0.12.0

# JSON
jsonlines>=3.1.0

dotenv

# Computer vision library
kornia>=0.6.0

# For SuperPoint keypoint detection
opencv-python>=4.5.0
transformers>=4.38.0

# -f https://download.pytorch.org/whl/torch_stable.html
# --extra-index-url https://download.pytorch.org/whl/cu121

opencv-python
tqdm
scipy
matplotlib
imgaug
torchscan
tensorboardX
datasets

argparse
scipy
opencv-contrib-python
imageio
tensorflow
pyyaml
jupyter
scikit-learn
torchgeometry
torchsummary
coloredlogs

# Google Earth Engine dependencies
earthengine-api>=0.1.384
geemap>=0.31.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0
python-dotenv>=1.0.0
retry>=0.9.2

# Google Maps API dependencies
requests>=2.28.0